FL-3797: Fix klibc failing to compile on PIE enabled compilers
Basically, the ebuild includes a version of the linux kernel,
which won't compile with pie enabled. Due to a bug, it doesn't
pick up the directions in the makefile to stop using pie while
compiling the kernel, so a special patch is needed.

Very unsure if Gentoo cares to fix this.