Add better sync-branch support,auto-sync-enforcing
sync-branch allows a repos.conf entry to specify which
branch to clone instead of the default master branch.

auto-sync-enforcing = yes enforces that the sync-uri
and sync-branch values in repos.conf entries match
those in the git repositories. E.g. to switch from
"master" to "1.0-prime" in the following example
it is sufficient to set sync-branch = 1.0-prime :

[DEFAULT]
main-repo = core-kit

[core-kit]
location = /var/git/core-kit
sync-type = git
sync-uri = https://github.com/funtoo/core-kit
sync-branch = master
auto-sync = yes
auto-sync-enforcing = yes

and run emerge --sync . If auto-sync-enforcing  is not
set (or is set to no) emerge --sync will only issue
the "git pull" command on the git repository.

And if the sync-uri is changed, emerge --sync will
redownload the git repository from the new sync-uri.

Currently, sync-branch and auto-sync-enforcing are only
supported by the git module.

The motivation for adding sync-branch and auto-sync-enforcing
is to simplify management of git branches which are essential
for kits to work properly. Switching branches by hand is rather
error-prone, but remains an option if auto-sync-enforcing = no is
set.

The reason for including branch support management in the portage
codebase instead of splitting it out into a separate tool is to
maintain a single point of truth in the system: the repos.conf entries
should be authorative for repository management. If you copy over
/etc to a new system and run emerge --sync it should work, instead of
having to rely on an external application to recreate the repositories
on the new system.