FL-3809 FL-3815: Add joomlatools-console
Joomlatools-console is a command line tool to
manage all aspects of joomla installation.
Joomlatools-console should replace the joomla
ebuild which is far more limited at managing
joomla.

According to the tools philosophy joomlatools-console
should be prefered. The joomla ebuild is hard to maintain
due to CVEs and is usually severely outdated.

joomlatools-console has been tested to be working with
php 5.6, 7.0 and 7.1. I haven't done extensive testing,
but from the looks of it does its job. I installed a
joomla site once on php 7.1 and it got me to the installation
page. I wish I could do more testing, but due to time constraints
I'll trust that the joomlatools-console developers did their job
well.