OK, some overhaul of meta-repo and metadata. First, I am trying to keep kit metadata as similar as possible, so I got rid of major changes to the metadata that turned out to be unnecessary.
There are still minor differences, and we
have a new metadata/version.json which defines the version of metadata
in the meta-repo. 10 == 1.3 release. <10 == 1.2 release or earlier.

And also -- we now will have different branches for meta-repo!
master branch is assumed to be release-1.2 or 1.0. For 1.3+, the
meta-repo branch will be "1.3-release", etc. This allows our metadata
to just describe what is in the actual current release, rather than
try to specify what is in other releases. This, in turn, cleans up
the metadata generation code by quite a bit!

Now, the user still specifies the release in /etc/ego.conf's release
variable, as "1.2", "1.3", etc. but ego will now know to look for "1.0"
or "1.2" in master, and "1.3" in the new branch. And we also make the
default release "1.2" if no default release is defined in the metadata
itself.

IMPORTANT!!!!

All this code needs to be tested. There are likely some problems. But
it goes a long way towards solving a lot of the complexity regarding our
overhaul of kits for 1.3. The overall concept is sound -- just needs
testing and validation.

The end-result should be that in meta-repo master, we have metadata
in 1.2-format, and in 1.3-release, we can have metadata in a slightly
updated format. And merge-scripts should generate both the old metadata
and the new, depending on which release they are generating for. So they
will put the old metadata in master, and the newer in 1.3-release.