FL-6635: Added newest version of net-im/slack-bin (4.1.1).
The current one (3.3.3) could not start due to segmentation fault.
The newest with major verion 3 (3.4.2) did work, but warns about it being old, and asking the user to upgrade.

The ebuild is taken from Gentoo, but with a single line added for setting the correct permissins (4755) on /opt/slack/chrome-sandbox, as it otherwise starts up with error: [6885:1030/191119.808429:FATAL:setuid_sandbox_host.cc(157)] The SUID sandbox helper binary was found, but is not configured correctly. Rather than run without sandboxing I'm aborting now. You need to make sure that /opt/slack/chrome-sandbox is owned by root and has mode 4755.