FL-6739, FL-6736: Remove LD_LIBRARY_PATH and GTK_THEME env variables
FL-6739: I have removed the LD_LIBRARY_PATH variable from the wrapper
script of thunderbird-bin.  This environment variable was causing
Firefox to crash when clicking a hyperlink in Thunderbird.

I purposefully am using a wrapper script instead of a symlink because
I have a hunch that thunderbird is looking in its dirname for helper
libraries.  I think the Freedesktop stuff might be happier with the
real path to the binary, too.

This fixes at a minimum the hyperlink side of FL-6739, possibly the
entire bug.

FL-6736: This patch undoes the workaround applied in FL-6736 because
there was a fix that went into upstream that addresses some of the
weirdness with Adwaita-dark.  If GTK_THEME=Adwaita is still required,
I can insert the workaround again.