Update app-editor/emacs to 26.3
This is an update long overdue.  I finally got around to cleaning up
and testing some of the acct-group madness.

+ Changes and minor fixes have been pulled from Gentoo.

+ I have opted to keep the older app-emacs/emacs-common-gentoo catpkg
provided by Funtoo’s Gentoo snapshots, which uses the older
‘enewgroup’ function instead of the newer acct-group/* virtual
packages.  This reduces the footprint in kit-fixups for now.

I have tested this in containers running Funtoo 1.3 and Funtoo 1.4.
This package emerges on both systems, even if the games USE flag is
set.