FL-6739: Remove injected environment variables from firefox-bin
These environment variables have the potential to cause the same
issues we saw in thunderbird-bin when opening other applications from
Firefox/Thunderbird.

See this comment and the following comment regarding the wrapper
script for thunderbird-bin:

https://bugs.funtoo.org/browse/FL-6739?focusedCommentId=43102&page=com.atlassian.jira.plugin.system.issuetabpanels:comment-tabpanel#comment-43102

I can’t find references to a requirement of using LD_LIBRARY_PATH with
Mozilla’s precompiled builds.  So, I think these variables are
remnants of the past.

https://support.mozilla.org/en-US/kb/install-firefox-linux

I have only changed Firefox 70 because it is in a testing state in
Funtoo.  There is little use in fixing something that isn’t quite
broken in an older version of Firefox.