Retrieve latest postman version from changelog
This patch reworks the postman-bin autogen to fetch the release notes to
figure out what the latest version is, instead of figuring it out by
checking the filename of the "latest" download link. It turns out that
postman developers aren't doing a great job with keeping the "latest"
download link... latest. According to their website, the latest version
for Linux is 8.12.4, when in reality it's 9.1.3. This should resolve our
postman versioning issues for now.