Add entities catalog to manage gid/uid
The entities defined are based on  the tool
https://github.com/geaaru/entities/
but could be easy used with others tools and/or
with a review of the users.eclass.

Hereinafter, the list of the files and a
description of the mission:

* `groups.yaml`: contains a map of all groups with
  static gid availables. If a group has also a related
   user it's used the `user` attribute.

* `users.yaml` contains a map of all users with
   static uid that are without a related group.

* `groups-dynamic.yaml`: contains a map of all groups
  with dynamic gid. In general, it's better to allocate a static
  uid/gid for a new service but it's not always the right solution
  for all use cases.

* `groups-dynamic.yaml`: contains a map of all groups
  with dynamic gid.

* `users-dynamic.yaml`: contains a map of all usrs
  with dynamic uid without a related group.

I left a "to check" comment where i haven't validate directly
the configuration of the specific service.

In general, by default I defined a shell /sbin/nologin and
a /dev/null homedir.