This is a fix for dev-python/pytz.
It was previously using the template-based pypi-1 generator. This generator
is not as robust as the pypi-simple-1 generator. In this specific case, we
needed the generator to find the lastest release with an sdist. The pypi-1
just grabs the latest release and fails if there is no source distribution,
and this was impacting the 2022.5 release.