Multiple updates related to kernel 6.4:
nvidia-drivers-535.86.05 -- this is a work in progress. There are
minor errors with the ebuild (not affecting functionality.) One
cosmetic error is that it looks like the 6.4 kernel objtool may
fix this, but we will potentially still see errorsi like this:
https://lore.kernel.org/lkml/202203311730.Io171QCN-lkp@intel.com/T/

Update zfs, zfs-kmod to autogen the upstream 2.2.99 release
(which is a bogus upstream tag representing "master" but this
version does work with kernel 6.4.)

New debian-sources-6.4.4_p1 which is being tested with the above.

I experienced a problem where the first boot didn't work, and
modules did not load. Presumably because depmod -a was not run.
This could be related to the nvidia-kernel-modules ebuild as
I believe it was run last.