Update pypi-compat-1 generator:
* Automatically detect when our S= variable may not be correct due
  to ``_`` vs ``-`` and adjust automatically.

* Allow ``du_pep517`` to be set to ``generator`` which will then
  tell the generator to auto-detect the ``du_pep517`` setting in
  the package by looking at ``pyproject.toml``. This currently
  will recognize: flit, hatchling, setuptools, and also deal
  with adding setuptools_scm as a build dependency if detected.

This is being tested with sphinxcontrib ebuilds which are moving
to pure auto-detected of du_pep517.

This commit also includes clean-up of all ancient sphinxcontrib
generators which used templates, and removes "autobahn" autogen
and cleans up a lot of unused sphinxcontrib autogens which we
shouldn't have needed in the first place.