Various improvements (WIP, need testing) to python-compat-1 generator.
It's now possible to specify extensions: cargo (previously
it was turned on automatically if cargo was in inherit --
this is still supported for backwards compatibility.)

Also supports extensions: local-only to use all the features
of pypi-compat-1 that are relevant if you happen to have an
sdist locally you want to use, or if you for some reason
need distutils-r1 but don't have any things to fetch from pypi.

Requires latest metatools git master.