[net-kit] net-misc/networkmanager Introduced 'jansson' dependency and removed 'dhcpcanon' flag from NetworkManager
This update includes the addition of 'jansson' to the dependencies of the NetworkManager, enhancing its features. It also includes the removal of the 'dhcpcanon' flag from the Meson configurations for NetworkManager as it is no longer necessary, thereby simplifying the configuration. These modifications aim to improve compatibility and streamline the feature set.