[eclass] go-module: Review logic of EGO_BUNDLE_POSTFIX
The EGO_BUNDLE_POSTFIX variable now use `-` at the
end in order to support mark golang bundle without
sha. For example, atm the forgejo project is autogen
with a tarball downloaded with builtin-dirlisting
generator and the tag sha is not available.

In this case in the ebuild we need to override default
EGO_BUNDLE_POSTFIX value and drop the final `-`
to correctly handle the mark bundle.